# back/swagger.json
{
  "basePath": "/",
  "consumes": ["application/json"],
  "definitions": {
    "IComment": {
      "properties": {
        "_id": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "value": {
          "type": "string"
        },
        "creation": {
          "type": "date"
        }
      },
      "required": ["userName", "value"],
      "type": "object"
    },
    "IPost": {
      "properties": {
        "_id": {
          "type": "string"
        },
        "format": {
          "type": "string"
        },
        "creator": {
          "type": "string"
        },
        "likes": {
          "userName": {
            "type": "string"
          },
          "value": {
            "type": "number"
          }
        },
        "tags": {
          "type": "array<string>"
        },
        "visibility": {
          "type": "string"
        },
        "caption": {
          "type": "string"
        },
        "comments": "array<String>"
      },
      "required": ["_id", "visibility", "creator", "format"],
      "type": "object"
    },
    "IUser": {
      "properties": {
        "_id": {
          "type": "string"
        },
        "role": {
          "type": "string"
        },
        "login": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "friends": {
          "type": "array<String>"
        }
      },
      "required": ["_id", "login", "password"],
      "type": "object"
    }
  },
  "info": {
    "title": "DragonDrop",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    }
  },
  "paths": {
    "/post": {
      "get": {
        "operationId": "GetAllpost",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/IPost"
              }
            }
          }
        },
        "security": [],
        "parameters": []
      },
      "post": {
        "operationId": "createPost",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "No content"
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "properties": {
                "_id": {
                  "type": "string"
                },
                "format": {
                  "type": "string"
                },
                "creator": {
                  "type": "string"
                },
                "likes": {
                  "userName": {
                    "type": "string"
                  },
                  "value": {
                    "type": "number"
                  }
                },
                "tags": {
                  "type": "array<string>"
                },
                "visibility": {
                  "type": "string"
                },
                "caption": {
                  "type": "string"
                },
                "comments": [
                  {
                    "writer": {
                      "type": "string"
                    },
                    "value": {
                      "type": "string"
                    },
                    "creation": {
                      "type": "date"
                    }
                  }
                ]
              },
              "required": ["_id", "login", "password"],
              "type": "object",

              "title": "CreateBody"
            }
          }
        ]
      }
    },
    "/post/id": {
      "get": {
        "operationId": "getPostId",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "array",
              "items": {
                "type": "number"
              }
            }
          }
        },
        "security": []
      }
    },
    "/post/{id}": {
      "get": {
        "operationId": "getPostById",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/IPost"
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "value": "601c5d9c2ea92508445002ef"
          }
        ]
      }
    },
    "/post/comment/{idPost}/{idComment}": {
      "delete": {
        "operationId": "deleteComment",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/IPost"
              }
            }
          }
        },
        "security": []
      }
    },

    "/post/comment/{id}": {
      "put": {
        "operationId": "addComment",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "userName": { "type": "string" },
              "value": { "type": "string" }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/IComment"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "value": "601c5d9c2ea92508445002ef"
          }
        ]
      }
    },
    "/post/like/{id}": {
      "put": {
        "operationId": "addLike",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "No content"
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "properties": {
                "userName": {
                  "type": "string"
                },
                "likeValue": {
                  "type": "number"
                }
              },
              "type": "object",
              "title": "addLike"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "value": "601c5d9c2ea92508445002ef"
          }
        ]
      }
    },
    "/user": {
      "get": {
        "operationId": "GetAllUser",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/IUser"
              }
            }
          }
        },
        "security": [],
        "parameters": []
      },
      "post": {
        "operationId": "CreateUser",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "No content"
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "properties": {
                "_id": {
                  "type": "string"
                },
                "role": {
                  "type": "string"
                },
                "login": {
                  "type": "string"
                },
                "password": {
                  "type": "string"
                },
                "friends": {
                  "type": "array<String>"
                }
              },
              "required": ["_id", "login", "password"],
              "type": "object",

              "title": "CreateBody"
            }
          }
        ]
      }
    },
    "/user/{id}": {
      "get": {
        "operationId": "getUserById",
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/IUser"
              }
            }
          }
        },
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "value": "6017e245da8a8046e41a2123"
          }
        ]
      }
    }
  },
  "produces": ["application/json"],
  "swagger": "2.0",
  "securityDefinitions": {}
}